Code and whitespace cleanup
Widget names have been converted from CamelCase to snake_case. A larger warning
is displayed in stderr if the config file can't be found or is invalid. The
default theme uri has been fixed. The config file has been renamed to
"config.def.json" to indicate that this is the default config.

Ref #29