waf: update object and library compilation and installation
curl/dbus objects are now correctly compiled with -fPIC for linking with
the shared objects. Shared objects are installed in LIBDIR, which is
configured as ${PREFIX}/lib/wkline by default.

LIBDIR can be overridden for running wkline with another library search
path, this can be useful for widget development/usage of wkline without
installing it in a system path:

    cd /home/user/wkline
    waf clean configure build --libdir='/home/user/wkline/out/usr/local/lib/wkline'
    waf install --destdir='out'

TODO: find a better way to load libraries with a relative search path.