config: refactor config getter functions
Instead of returning a json_t that needs to be handled,
wkline_get_config and widget_get_config now accept a void pointer that
will receive the config value. json_typeof() is used to detect the type
of the config value, and the appropriate json_*_value() function is used
to assign the value to the value pointer.

The *_get_config() functions now return an int error code if they fail,
which can be handled if necessary.

Ref #42