Add the mpris widget
The mpris widget displays the artist and title for any media player that
implements the MPRIS DBus Interface Specification such as vlc, spotify,
mplayer, and others.

For more information on MPRIS:

http://specifications.freedesktop.org/mpris-spec/latest/

Pass "player_name" to the config to connect to
org.mpris.MediaPlayer2.[PLAYER] on your DBus session. If no player is
passed, the widget will choose the first mpris-enabled player that
becomes available to get metadata from.