widgets: add timeout for threads to finish
If threads don't send the condition signal (with the clean exit macro)
before the cond wait times out the thread will be cancelled; threads
that don't reach the epoll wait in their main loop before being
signalled to exit may block until the next timed loop, which can be very
long for e.g. e-mail and IP widgets.

Currently the timeout is two seconds, which should be enough for most
widgets to finish.