widgets/now_playing_mpris: rewrite
Improve render quirks and support configuration-free mpris widget
suitable for use with multiple media players.

The widget polls for mpris players. When it finds one, it subscribes to
events and updates the player status appropriately.

If no player name is configured, when the player exits, polling starts
again for another player.