core: Use correct screen Y offset when positioning bar
When using offset monitors the bar would incorrectly be positioned
outside of the monitor. When offsetting a monitor by e.g. +0+100, the
coordinates of the upper left corner of the monitor are (0, 100) and not
(0, 0) which was previously assumed when setting the bar position.

This update fixes the issue by considering *both* the `dest.x` and
`dest.y` coordinates from `gdk_screen_get_monitor_geometry()` as the
upper left corner of the monitor.